+++ backend/changes.md
@@ -1,6 +1,11 @@
 # Change Log
 
+## Release v2.2.2
+__LTS 30/08/2021__
+
+ - Adicionando endpoing `DELETE /account/:a/user/:u`
+
 ## Release v2.2.1
 __LTS 22/06/2021__
 
-- Adicionando rota '/user/password/force'
+ - Adicionando rota '/user/password/force'